Widen literal pools to u16 and lift every per-function capacity ceiling
Closes #44, #45 and #63. All four capacity repros now compile and run where
each previously failed at compile time; reference Lua accepts all of them.

#44 had four independent ceilings, all consequences of literal indices being
8-bit operands:

  {1, ...} x300              too many fields in table constructor
  {x1=1, ... x300=300}       too many literal strings
  t.f1 = 1 ... t.f300 = 300  too many literal strings
  300 distinct number literals   too many literal numbers

Array constructors now flush in batches: SET_LIST carries a batch ordinal in
its previously reserved Bx and inserts from batch * 255 + 1, since the old
instruction always restarted at key 1. That needed no encoding change.

The literal pools move to 16-bit ids carried in Bx, which does change the
instruction encoding. PUSH_NUM, PUSH_STRING and SET_GLOBAL had B and C free.
GET_GLOBAL, GET_FIELD and SET_FIELD did not - their words were already full -
so their inline-cache slots narrow to 8 bits with 255 meaning "uncached",
matching the sentinel scheme adopted for SET_FIELD in the same change. Cached
sites are capped at 255 per function and the rest take the slow path; measured
at 1.81x on a map-backed table, about 18ns per write, and instruction cost is
unchanged either way. Nothing in the repo comes close to the cap: the most
field-heavy example has 83 field instructions in the entire file, and the most
global-heavy has ten distinct names that are all builtins needing no slot.

SET_FIELD could not simply swap operands because A carries the receiver's
stack offset, which is nonzero only for earlier destinations in a multiple
assignment. It splits into SET_FIELD (cached, implied offset zero) and
SET_FIELD_AT (uncached, explicit offset); the parser emits the latter and
finalization rewrites offset-zero sites to the former. The hot path keeps its
cache and the rare multi-lvalue path pays instead of everything paying.

INIT_FIELD_PINNED widens only its key: a small templated constructor late in a
function can have every key above 255, but the template entry index cannot,
because templates are an optional optimization with an existing fallback.
Constructors past 255 entries now skip the template rather than failing.

Table templates become Vec<Vec<u16>> - they hold string-pool ids, so leaving
them u8 would have kept a hidden ceiling behind the widened instructions.

Pools are u16 rather than u32 deliberately. 65536 is 200x the motivating cases,
and every call already interns all of a function's string literals into the
State, so that many distinct literals is already an extreme workload; u32 would
require multiword encoding and a different literal-loading design.

#45: add_local grew num_locals whenever locals.len() exceeded it, but
parameters were pushed without updating it, so sibling scopes in a function
with parameters re-triggered growth and every call pushed extra nils. It now
maximizes locals.len() - num_params.

#63: a call spanning several lines was attributed to its closing line, because
OP_CALL is emitted after the argument list is parsed. The opening line is now
carried on PrefixExp::FunctionCall and used at emission, so tracebacks blame
the same line reference does. The line and argument count are packed into four
bytes to keep that variant from dominating an enum the parser copies
constantly.

Oversized GET_FIELD sites raised InternalError, which since the error-taxonomy
change means "VM bug"; it is now a line-bearing SyntaxError, consistent with
the SET_FIELD path beside it.

Verification treats sentinel 255 as uncached for all three cache-bearing
opcodes, accepts it anywhere rather than only after the slots are exhausted,
excludes it from declared counts, and still requires real slots to be
sequential. Every Bx literal and every u16 template key is bounds-checked
before use, with a forged-snapshot test per opcode sharing the string-literal
arm so a future edit cannot silently drop one.

Breaking change: snapshot format v4 to v5, rejected by the existing strict
equality gate with no fallback parsing. The golden fixture is regenerated.

Coded by an LLM.

+++ src/vm/eval_index.rs
@@ -16,8 +16,8 @@ impl State {
     pub(super) fn instr_get_field(
         &mut self,
         frame: &mut Frame,
-        field_id: u8,
-        cache_idx: u16,
+        field_id: u16,
+        cache_idx: u8,
         local_cost: &mut u64,
     ) -> Result<()> {
         // Pop value, handle both tables and strings
@@ -373,8 +373,8 @@ impl State {
     pub(super) fn instr_get_global(
         &mut self,
         frame: &Frame,
-        string_num: u8,
-        cache_idx: u16,
+        string_num: u16,
+        cache_idx: u8,
     ) -> Result<()> {
         let s = &frame.bytecode().string_literals[string_num as usize];
         let cache = frame.caches.global_lookup.get(cache_idx as usize);
